refactor: Request 객체 응집도를 고려하여 패키지 위치 변경
