+++ src/main/java/woowa/camp/jspcafe/repository/dto/request/ArticleUpdateRequest.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.repository.dto;
+package woowa.camp.jspcafe.repository.dto.request;
 
 public record ArticleUpdateRequest(String title, String content) {
 }
