+++ src/main/java/woowa/camp/jspcafe/repository/dto/request/UserUpdateRequest.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.repository.dto;
+package woowa.camp.jspcafe.repository.dto.request;
 
 public record UserUpdateRequest(String password, String nickname) {
 }
