+++ src/main/java/woowa/camp/jspcafe/repository/user/DBUserRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 import woowa.camp.jspcafe.domain.User;
 import woowa.camp.jspcafe.infra.DatabaseConnector;
-import woowa.camp.jspcafe.repository.dto.UserUpdateRequest;
+import woowa.camp.jspcafe.repository.dto.request.UserUpdateRequest;
 
 public class DBUserRepository implements UserRepository {
