+++ src/main/java/woowa/camp/jspcafe/repository/user/UserRepository.java
@@ -3,7 +3,7 @@ package woowa.camp.jspcafe.repository.user;
 import java.util.List;
 import java.util.Optional;
 import woowa.camp.jspcafe.domain.User;
-import woowa.camp.jspcafe.repository.dto.UserUpdateRequest;
+import woowa.camp.jspcafe.repository.dto.request.UserUpdateRequest;
 
 public interface UserRepository {
