+++ src/main/java/woowa/camp/jspcafe/service/UserService.java
@@ -7,8 +7,8 @@ import org.slf4j.LoggerFactory;
 import woowa.camp.jspcafe.domain.User;
 import woowa.camp.jspcafe.domain.exception.UserException;
 import woowa.camp.jspcafe.repository.user.UserRepository;
-import woowa.camp.jspcafe.repository.dto.UserUpdateRequest;
-import woowa.camp.jspcafe.service.dto.RegistrationRequest;
+import woowa.camp.jspcafe.repository.dto.request.UserUpdateRequest;
+import woowa.camp.jspcafe.service.dto.request.RegistrationRequest;
 import woowa.camp.jspcafe.repository.dto.response.UserResponse;
 import woowa.camp.jspcafe.infra.time.DateTimeProvider;
