+++ src/main/java/woowa/camp/jspcafe/service/dto/request/ArticleWriteRequest.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.service.dto;
+package woowa.camp.jspcafe.service.dto.request;
 
 public record ArticleWriteRequest(Long authorId, String title, String content) {
 }
