+++ src/main/java/woowa/camp/jspcafe/service/dto/request/RegistrationRequest.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.service.dto;
+package woowa.camp.jspcafe.service.dto.request;
 
 public record RegistrationRequest(String email, String nickname, String password) {
 }
