+++ src/main/java/woowa/camp/jspcafe/service/dto/request/ReplyWriteRequest.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.service.dto;
+package woowa.camp.jspcafe.service.dto.request;
 
 public record ReplyWriteRequest(Long userId, Long articleId, String content) {
 }
