+++ src/main/java/woowa/camp/jspcafe/web/servlet/user/UserEditServlet.java
@@ -14,7 +14,7 @@ import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import woowa.camp.jspcafe.domain.User;
-import woowa.camp.jspcafe.repository.dto.UserUpdateRequest;
+import woowa.camp.jspcafe.repository.dto.request.UserUpdateRequest;
 import woowa.camp.jspcafe.service.UserService;
 import woowa.camp.jspcafe.web.utils.PathVariableExtractor;
